+++ Changelog.md
@@ -2,12 +2,22 @@
 
 ### Ignition Gazebo 2.X.X
 
+1. Use scene ambient and background color information in sensor
+   configuration. 
+    * [Pull Request 268](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/268)
+
 1. Added an SDF message to the start of log files.
     * [Pull Request 257](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/257)
 
 ## Ignition Gazebo 1.x
 
 ### Ignition Gazebo 1.X.X
+
+1. Added LiftDragPlugin (ported from Gazebo classic)
+    * [Pull Request 256](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/256)
+
+1. Added test for log record and playback.
+    * [Pull Request 263](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/263)
 
 1. More ign-msgs <-> SDF conversions: Inertial, Geometry, Material
     * [Pull Request 251](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/251)